Improve Audit result table views
